+++ functions/package.json
@@ -1,35 +1,36 @@
 {
-  "name": "functions",
-  "scripts": {
-    "build": "tsc",
-    "serve": "npm run build && firebase serve --only functions",
-    "shell": "npm run build && firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log",
-    "copyMessageTranslations": "node scripts/copyPublicMessageToFunctions.js"
-  },
-  "engines": {
-    "node": "8"
-  },
-  "main": "lib/index.js",
-  "dependencies": {
-    "@sendgrid/mail": "^6.5.5",
-    "cors": "^2.8.5",
-    "firebase-admin": "^8.6.0",
-    "firebase-functions": "^3.3.0",
-    "libphonenumber-js": "^1.7.48",
-    "lodash": "^4.17.15",
-    "node-fetch": "^2.6.0",
-    "nodemailer": "^6.4.6",
-    "ovh": "^2.0.3"
-  },
-  "devDependencies": {
-    "@types/cors": "^2.8.6",
-    "@types/node-fetch": "^2.5.5",
-    "@types/nodemailer": "^6.4.0",
-    "firebase-functions-test": "^0.1.6",
-    "typescript": "^3.2.2"
-  },
-  "private": true
+    "name": "functions",
+    "scripts": {
+        "build": "tsc",
+        "serve": "npm run build && firebase serve --only functions",
+        "shell": "npm run build && firebase functions:shell",
+        "start": "npm run shell",
+        "deploy": "firebase deploy --only functions",
+        "logs": "firebase functions:log",
+        "copyMessageTranslations": "node scripts/copyPublicMessageToFunctions.js"
+    },
+    "engines": {
+        "node": "8"
+    },
+    "main": "lib/index.js",
+    "dependencies": {
+        "@sendgrid/mail": "^6.5.5",
+        "cors": "^2.8.5",
+        "firebase-admin": "^8.6.0",
+        "firebase-functions": "^3.3.0",
+        "libphonenumber-js": "^1.7.48",
+        "lodash": "^4.17.15",
+        "node-fetch": "^2.6.0",
+        "nodemailer": "^6.4.6",
+        "ovh": "^2.0.3",
+        "twilio": "^3.42.2"
+    },
+    "devDependencies": {
+        "@types/cors": "^2.8.6",
+        "@types/node-fetch": "^2.5.5",
+        "@types/nodemailer": "^6.4.0",
+        "firebase-functions-test": "^0.1.6",
+        "typescript": "^3.2.2"
+    },
+    "private": true
 }
