+++ functions/src/callable/utils/sms/twilio.ts
@@ -1,0 +1,6 @@
+import { NotificationParams } from '../notification'
+
+export const sendSmsViaTwilio = async (
+    params: NotificationParams,
+    messageBody: string
+) => {}
